tests: Move hash_file() into tests/utils.py
