+++ monkey/tests/utils.py
@@ -1,5 +1,7 @@
 import ctypes
+import hashlib
 import os
+from pathlib import Path
 
 
 def is_user_admin():
@@ -7,3 +9,12 @@ def is_user_admin():
         return os.getuid() == 0
 
     return ctypes.windll.shell32.IsUserAnAdmin()
+
+
+def hash_file(filepath: Path):
+    sha256 = hashlib.sha256()
+    with open(filepath, "rb") as f:
+        for block in iter(lambda: f.read(65536), b""):
+            sha256.update(block)
+
+    return sha256.hexdigest()
